[ Widget Preview ] Fix type error when retrieving flags from persistent preferences
Also moves dtd_services_test.dart to actually use the
WidgetPreviewScaffoldDtdServices impelemntation, which would have caught
the type cast issue in the implementation.
