+++ packages/flutter_tools/pubspec.yaml
@@ -110,7 +110,7 @@ dependencies:
   term_glyph: 1.2.2
   typed_data: 1.4.0
   vm_service_interface: 2.0.1
-  watcher: 1.1.3
+  watcher: 1.1.4
   web: 1.1.1
   web_socket: 1.0.1
   yaml_edit: 2.2.2
@@ -127,4 +127,4 @@ dev_dependencies:
 dartdoc:
   # Exclude this package from the hosted API docs.
   nodoc: true
-# PUBSPEC CHECKSUM: 37t3vr
+# PUBSPEC CHECKSUM: ho7nke
