+++ packages/flutter_tools/test/widget_preview_scaffold.shard/widget_preview_scaffold/lib/src/dtd/dtd_services.dart
@@ -25,6 +25,7 @@ class WidgetPreviewScaffoldDtdServices with DtdEditorService {
   static const kResolveUri = 'resolveUri';
   static const kSetPreference = 'setPreference';
   static const kGetPreference = 'getPreference';
+  static const kGetDevToolsUri = 'getDevToolsUri';
 
   /// Error code for RpcException thrown when attempting to load a key from
   /// persistent preferences that doesn't have an entry.
@@ -92,12 +93,14 @@ class WidgetPreviewScaffoldDtdServices with DtdEditorService {
   /// preferences map.
   ///
   /// Returns null if [key] is not in the map.
-  Future<String?> getPreference(String key) async {
+  Future<Object?> getPreference(String key) async {
     try {
-      final response = StringResponse.fromDTDResponse(
-        (await _call(kGetPreference, params: {'key': key}))!,
-      );
-      return response.value;
+      final response = await _call(kGetPreference, params: {'key': key});
+      return switch (response!.type) {
+        'StringResponse' => StringResponse.fromDTDResponse(response).value,
+        'BoolResponse' => BoolResponse.fromDTDResponse(response).value,
+        _ => throw StateError('Unexpected response type: ${response.type}'),
+      };
     } on RpcException catch (e) {
       if (e.code == kNoValueForKey) {
         return null;
@@ -110,11 +113,11 @@ class WidgetPreviewScaffoldDtdServices with DtdEditorService {
   ///
   /// If [key] is not set, [defaultValue] is returned.
   Future<bool> getFlag(String key, {bool defaultValue = false}) async {
-    final result = await getPreference(key);
+    final result = await getPreference(key) as bool?;
     if (result == null) {
       return defaultValue;
     }
-    return bool.tryParse(result) ?? defaultValue;
+    return result;
   }
 
   /// Sets [key] to [value] in the persistent preferences map.
