+++ pubspec.yaml
@@ -133,7 +133,7 @@ dependencies:
   meta: 1.17.0
   metrics_center: 1.0.13
   mime: 2.0.0
-  native_toolchain_c: 0.17.1
+  native_toolchain_c: 0.17.2
   nested: 1.0.0
   node_preamble: 2.0.2
   package_config: 2.2.0
@@ -175,7 +175,7 @@ dependencies:
   test_core: 0.6.12
   typed_data: 1.4.0
   url_launcher: 6.3.2
-  url_launcher_android: 6.3.22
+  url_launcher_android: 6.3.23
   url_launcher_ios: 6.3.4
   url_launcher_linux: 3.2.1
   url_launcher_macos: 3.2.3
@@ -184,13 +184,13 @@ dependencies:
   url_launcher_windows: 3.1.4
   vector_math: 2.2.0
   video_player: 2.10.0
-  video_player_android: 2.8.13
+  video_player_android: 2.8.14
   video_player_avfoundation: 2.8.4
   video_player_platform_interface: 6.4.0
   video_player_web: 2.4.0
   vm_service: 15.0.2
   vm_snapshot_analysis: 0.7.6
-  watcher: 1.1.3
+  watcher: 1.1.4
   web: 1.1.1
   web_socket: 1.0.1
   web_socket_channel: 3.0.3
@@ -212,4 +212,4 @@ dependencies:
   pedantic: 1.11.1
   quiver: 3.2.2
   yaml_edit: 2.2.2
-# PUBSPEC CHECKSUM: 9hqic
+# PUBSPEC CHECKSUM: 5h4lnl
